Make Flashlight optional
